Update common attention
